requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu
ultralytics
opencv-python-headless
python-multipart
numpy
pillow